+++ src/main/java/org/example/catch_line/history/controller/HistoryController.java
@@ -6,9 +6,8 @@ import org.example.catch_line.common.SessionUtils;
 import org.example.catch_line.common.constant.Status;
 import org.example.catch_line.history.model.dto.HistoryResponse;
 import org.example.catch_line.history.service.HistoryService;
-import org.example.catch_line.reservation.service.ReservationService;
-import org.example.catch_line.waiting.service.WaitingService;
-import org.springframework.security.core.parameters.P;
+import org.example.catch_line.booking.reservation.service.ReservationService;
+import org.example.catch_line.booking.waiting.service.WaitingService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
