--- src/main/java/org/example/catch_line/reservation/validation/ValidReservationDate.java
@@ -1,20 +1,0 @@
-package org.example.catch_line.reservation.validation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-@Constraint(validatedBy = ReservationDateValidator.class)
-@Target({ElementType.METHOD, ElementType.FIELD, ElementType.ANNOTATION_TYPE, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ValidReservationDate {
-	String message() default "예약 날짜는 현재 날짜 다음날부터 7일 이내여야 합니다";
-
-	Class<?>[] groups() default {};
-
-	Class<? extends Payload>[] payload() default {};
-}
