+++ src/test/java/org/example/catch_line/reservation/service/ReservationServiceTest.java
@@ -7,6 +7,7 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import org.example.catch_line.booking.reservation.service.ReservationService;
 import org.example.catch_line.common.constant.Role;
 import org.example.catch_line.common.constant.SessionConst;
 import org.example.catch_line.common.constant.Status;
@@ -15,10 +16,10 @@ import org.example.catch_line.member.model.dto.SignUpRequest;
 import org.example.catch_line.member.model.entity.MemberEntity;
 import org.example.catch_line.member.repository.MemberRepository;
 import org.example.catch_line.member.service.AuthService;
-import org.example.catch_line.reservation.model.dto.ReservationRequest;
-import org.example.catch_line.reservation.model.dto.ReservationResponse;
-import org.example.catch_line.reservation.model.entity.ReservationEntity;
-import org.example.catch_line.reservation.repository.ReservationRepository;
+import org.example.catch_line.booking.reservation.model.dto.ReservationRequest;
+import org.example.catch_line.booking.reservation.model.dto.ReservationResponse;
+import org.example.catch_line.booking.reservation.model.entity.ReservationEntity;
+import org.example.catch_line.booking.reservation.repository.ReservationRepository;
 import org.example.catch_line.restaurant.model.dto.RestaurantCreateRequest;
 import org.example.catch_line.restaurant.model.dto.RestaurantResponse;
 import org.example.catch_line.restaurant.model.entity.constant.FoodType;
@@ -29,7 +30,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 
 import jakarta.servlet.http.HttpSession;
