+++ src/test/java/org/example/catch_line/waiting/service/WaitingServiceTest.java
@@ -7,6 +7,7 @@ import java.math.BigDecimal;
 import java.util.List;
 
 import org.assertj.core.api.Assertions;
+import org.example.catch_line.booking.waiting.service.WaitingService;
 import org.example.catch_line.common.constant.Role;
 import org.example.catch_line.common.constant.SessionConst;
 import org.example.catch_line.common.constant.Status;
@@ -19,11 +20,11 @@ import org.example.catch_line.restaurant.model.dto.RestaurantResponse;
 import org.example.catch_line.restaurant.model.entity.constant.FoodType;
 import org.example.catch_line.restaurant.model.entity.constant.ServiceType;
 import org.example.catch_line.restaurant.service.RestaurantService;
-import org.example.catch_line.waiting.model.dto.WaitingRequest;
-import org.example.catch_line.waiting.model.dto.WaitingResponse;
-import org.example.catch_line.waiting.model.entity.WaitingEntity;
-import org.example.catch_line.waiting.model.entity.WaitingType;
-import org.example.catch_line.waiting.repository.WaitingRepository;
+import org.example.catch_line.booking.waiting.model.dto.WaitingRequest;
+import org.example.catch_line.booking.waiting.model.dto.WaitingResponse;
+import org.example.catch_line.booking.waiting.model.entity.WaitingEntity;
+import org.example.catch_line.booking.waiting.model.entity.WaitingType;
+import org.example.catch_line.booking.waiting.repository.WaitingRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
